e2e_test/requirements_dev.txt
alembic==0.8.10
apache-airflow==1.10.2
Babel==2.6.0
bleach==2.1.4
cachetools==3.1.0
certifi==2018.11.29
chardet==3.0.4
Click==7.0
colorama==0.4.1
configparser==3.5.3
croniter==0.3.27
defusedxml==0.5.0
dill==0.2.9
docutils==0.14
enum34==1.1.6
Flask==0.12.4
Flask-Admin==1.5.2
Flask-AppBuilder==1.12.1
Flask-Babel==0.12.2
Flask-Caching==1.3.3
Flask-Login==0.4.1
Flask-OpenID==1.2.5
Flask-SQLAlchemy==2.3.2
flask-swagger==0.2.13
Flask-WTF==0.14.2
funcsigs==1.0.0
future==0.16.0
gitdb2==2.0.5
GitPython==2.1.11
google-api-core==1.7.0
google-api-python-client==1.7.8
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-cloud-tasks==0.4.0
googleapis-common-protos==1.5.8
grpc-google-iam-v1==0.11.4
grpcio==1.18.0
gunicorn==19.9.0
html5lib==1.0.1
httplib2==0.12.1
idna==2.8
iso8601==0.1.12
itsdangerous==1.1.0
Jinja2==2.10
json-merge-patch==0.2
lockfile==0.12.2
lxml==4.3.1
Mako==1.0.7
Markdown==2.6.11
MarkupSafe==1.1.0
monotonic==1.5
numpy==1.16.1
ordereddict==1.1
pandas==0.24.1
pendulum==1.4.4
protobuf==3.6.1
psutil==5.5.1
pyasn1==0.4.5
pyasn1-modules==0.2.4
Pygments==2.3.1
python-daemon==2.1.2
python-dateutil==2.8.0
python-editor==1.0.4
python-nvd3==0.15.0
python-slugify==2.0.1
python3-openid==3.1.0
pytz==2018.9
pytzdata==2018.9
PyYAML==3.13
requests==2.21.0
rsa==4.0
setproctitle==1.1.10
six==1.12.0
smmap2==2.0.5
SQLAlchemy==1.2.18
tabulate==0.8.2
tenacity==4.8.0
thrift==0.11.0
tzlocal==1.5.1
unicodecsv==0.14.1
Unidecode==1.0.23
uritemplate==3.0.0
urllib3==1.24.1
webencodings==0.5.1
Werkzeug==0.14.1
WTForms==2.2.1
zope.deprecation==4.4.0
